reduce jank for tipping
